@@ -12,13 +12,20 @@ plugins {
     id 'java'
 }
 
+apply plugin: 'idea'
+idea {
+    module {
+        inheritOutputDirs = true
+    }
+}
+
 apply plugin: 'io.codearte.nexus-staging'
 
 group 'com.uddernetworks.newocr'
 archivesBaseName = "NewOCR"
 version '1.1.1'
 
-sourceCompatibility = 1.8
+sourceCompatibility = 11
 
 repositories {
     mavenCentral()
@@ -36,11 +43,16 @@ dependencies {
 
     // https://mvnrepository.com/artifact/org.hsqldb/hsqldb
     compile group: 'org.hsqldb', name: 'hsqldb', version: '2.4.1'
+
+    // https://mvnrepository.com/artifact/it.unimi.dsi/fastutil
+    compile group: 'it.unimi.dsi', name: 'fastutil', version: '8.2.2'
 }
 
 nexusStaging {
-    username = ossrhUser
-    password = ossrhPassword
+    if (project.hasProperty("ossrhUser") && project.hasProperty("ossrhPassword")) {
+        username = ossrhUser
+        password = ossrhPassword
+    }
 }
 
 javadoc {
@@ -76,10 +88,14 @@ allprojects {
             mavenDeployer {
                 // POM signature
                 beforeDeployment { MavenDeployment deployment -> signing.signPom(deployment) }
-                // Target repository
-                repository(url: "https://oss.sonatype.org/service/local/staging/deploy/maven2/") {
-                    authentication(userName: ossrhUser, password: ossrhPassword)
+
+                if (project.hasProperty("ossrhUser") && project.hasProperty("ossrhPassword")) {
+                    // Target repository
+                    repository(url: "https://oss.sonatype.org/service/local/staging/deploy/maven2/") {
+                        authentication(userName: ossrhUser, password: ossrhPassword)
+                    }
                 }
+
                 pom.project {
                     name 'NewOCR'
                     description 'NewOCR is a library for simple but efficient OCR detection in pure Java.'
