@@ -4,11 +4,13 @@ package com.uddernetworks.newocr;
  * An object storing the data for a set of bounds (Both upper and lower)
  */
 public class FontBounds {
+    
     private int minFont;
     private int maxFont;
 
     /**
-     * Creates a FontBounds object from the given bounds
+     * Creates a FontBounds object from the given bounds.
+     *
      * @param minFont The upper limit of allowed font size
      * @param maxFont The lower limit of allowed font size
      */
@@ -18,7 +20,8 @@ public class FontBounds {
     }
 
     /**
-     * Gets the minimum allowed font size
+     * Gets the minimum allowed font size.
+     *
      * @return The minimum allowed font size
      */
     public int getMinFont() {
@@ -26,7 +29,8 @@ public class FontBounds {
     }
 
     /**
-     * Gets the maximum allowed font size
+     * Gets the maximum allowed font size.
+     *
      * @return The maximum allowed font size
      */
     public int getMaxFont() {
@@ -36,6 +40,7 @@ public class FontBounds {
     /**
      * Gets if the given size is contained or included in the font bounds. This method is simple but is good for a
      * visual comparison.
+     *
      * @param font The font size to compare
      * @return If the font is between or in the borders
      */
@@ -47,4 +52,5 @@ public class FontBounds {
     public String toString() {
         return "FontBounds[" + this.minFont + " - " + this.maxFont + "]";
     }
+    
 }
