@@ -1,5 +1,7 @@
 package com.uddernetworks.newocr;
 
+import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
+import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
 import java.util.Arrays;
 import java.util.Optional;
 
@@ -7,6 +9,7 @@ import java.util.Optional;
  * Meta for letters that is inserted into the database to help distinguish characters.
  */
 public enum LetterMeta {
+    
     NONE(0),
     EVEN_DOTS(1), // : =
     DOT_UNDER(2), // ! ?
@@ -14,6 +17,12 @@ public enum LetterMeta {
     PERCENT(4),   // %
     QUOTE(5);     // "
 
+    private static final Int2ObjectMap<LetterMeta> LETTER_META = new Int2ObjectOpenHashMap<>(values().length);
+    
+    static {
+        Arrays.stream(values()).forEach(value -> LETTER_META.put(value.id, value));
+    }
+    
     private int id;
 
     LetterMeta(int id) {
@@ -21,7 +30,8 @@ public enum LetterMeta {
     }
 
     /**
-     * Gets the LetterMeta's ID
+     * Gets the LetterMeta's ID.
+     *
      * @return The ID
      */
     public int getID() {
@@ -30,10 +40,12 @@ public enum LetterMeta {
 
     /**
      * Gets a {@link LetterMeta} from the given ID.
+     *
      * @param id The ID to get
      * @return The {@link LetterMeta} with the same ID as the one given
      */
     public static Optional<LetterMeta> fromID(int id) {
-        return Arrays.stream(values()).filter(letterMeta -> letterMeta.id == id).findFirst();
+        return Optional.ofNullable(LETTER_META.get(id));
     }
+    
 }
