@@ -1,13 +1,11 @@
 package com.uddernetworks.newocr;
 
 import com.uddernetworks.newocr.character.ImageLetter;
-
+import it.unimi.dsi.fastutil.ints.Int2ObjectLinkedOpenHashMap;
+import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import java.awt.image.BufferedImage;
 import java.io.File;
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.ExecutionException;
 
 /**
@@ -16,8 +14,10 @@ import java.util.concurrent.ExecutionException;
 public class ScannedImage {
 
     private transient File originalFile;
+    
     private transient BufferedImage originalImage;
-    private Map<Integer, List<ImageLetter>> grid = new LinkedHashMap<>();
+    
+    private final Int2ObjectMap<List<ImageLetter>> grid = new Int2ObjectLinkedOpenHashMap<>();
 
     public ScannedImage(File originalFile, BufferedImage originalImage) {
         this.originalFile = originalFile;
@@ -49,10 +49,13 @@ public class ScannedImage {
      */
     public int getFirstFontSize(OCRHandle ocrHandle) {
         try {
-            for (Map.Entry<Integer, List<ImageLetter>> entry : grid.entrySet()) {
-                for (ImageLetter imageLetter : entry.getValue()) {
+            for (var imageLetters : grid.values()) {
+                for (var imageLetter : imageLetters) {
                     int size = ocrHandle.getFontSize(imageLetter).get();
-                    if (size != -1) return size;
+                    
+                    if (size != -1) {
+                        return size;
+                    }
                 }
             }
         } catch (InterruptedException | ExecutionException e) {
@@ -64,23 +67,26 @@ public class ScannedImage {
 
     /**
      * gets the amount of lines in the image.
+     *
      * @return The amount of lines in the image
      */
     public int getLineCount() {
-        return this.grid.size();
+        return grid.size();
     }
 
     /**
      * Returns the raw, mutable grid of {@link ImageLetter}s internally used with the key of the mpa being the exact Y
      * position of the line.
+     *
      * @return The raw, mutable grid of values
      */
-    public Map<Integer, List<ImageLetter>> getGrid() {
-        return this.grid;
+    public Int2ObjectMap<List<ImageLetter>> getGrid() {
+        return grid;
     }
 
     /**
-     * Adds a line containing {@link ImageLetter}s
+     * Adds a line containing {@link ImageLetter}s.
+     *
      * @param y The exact Y position of the line
      * @param databaseCharacterList A list of {@link ImageLetter}s as the line
      */
@@ -90,20 +96,22 @@ public class ScannedImage {
 
     /**
      * Gets the line at the given index value.
+     *
      * @param y The index of the line
      * @return The line at the given index value
      */
     public List<ImageLetter> getLine(int y) {
-        return grid.get(new ArrayList<>(grid.keySet()).get(y));
+        return grid.values().stream().skip(y).findFirst().orElse(null);
     }
 
     /**
      * Gets both the line Y and values at the given index value.
+     *
      * @param y The index of the line
      * @return The line at the given Y index value
      */
-    public Map.Entry<Integer, List<ImageLetter>> getLineEntry(int y) {
-        return new ArrayList<>(grid.entrySet()).get(y);
+    public Int2ObjectMap.Entry<List<ImageLetter>> getLineEntry(int y) {
+        return grid.int2ObjectEntrySet().stream().skip(y).findFirst().orElse(null);
     }
 
     /**
@@ -123,4 +131,5 @@ public class ScannedImage {
     public File getOriginalFile() {
         return originalFile;
     }
+    
 }
