@@ -2,7 +2,8 @@ package com.uddernetworks.newocr.train;
 
 import com.uddernetworks.newocr.LetterMeta;
 import com.uddernetworks.newocr.character.SearchCharacter;
-
+import it.unimi.dsi.fastutil.doubles.DoubleArrayList;
+import it.unimi.dsi.fastutil.doubles.DoubleList;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.OptionalDouble;
@@ -22,13 +23,13 @@ public class TrainedCharacterData {
     private double sizeRatio = -1; //        Width / Height
     private boolean empty = true;
     private LetterMeta letterMeta = LetterMeta.NONE;
-
     private List<double[]> recalculatingList = new ArrayList<>();
-    private List<Double> recalculatingWidths = new ArrayList<>();
-    private List<Double> recalculatingHeights = new ArrayList<>();
+    private DoubleList recalculatingWidths = new DoubleArrayList();
+    private DoubleList recalculatingHeights = new DoubleArrayList();
 
     /**
-     * Creates a {@link TrainedCharacterData} from a character value
+     * Creates a {@link TrainedCharacterData} from a character value.
+     *
      * @param value The known character value
      */
     public TrainedCharacterData(char value) {
@@ -36,7 +37,8 @@ public class TrainedCharacterData {
     }
 
     /**
-     * Gets the assigned character value
+     * Gets the assigned character value.
+     *
      * @return The assigned character value
      */
     public char getValue() {
@@ -45,31 +47,35 @@ public class TrainedCharacterData {
 
     /**
      * Gets the calculated array of the percentages for each section. The array is always 16 elements with everything
-     * being <= 1
+     * being <= 1.
+     *
      * @return An array of percentages
      */
     public double[] getSegmentPercentages() {
-        return this.segmentPercentages;
+        return segmentPercentages;
     }
 
     /**
      * Gets if the trained character has a dot.
+     *
      * @return If the trained character has a dot
      */
     public boolean hasDot() {
-        return this.hasDot;
+        return hasDot;
     }
 
     /**
      * Sets if the trained character has a dot.
+     *
      * @param hasDot If the trained character has a dot
      */
     public void setHasDot(boolean hasDot) {
-        if (!this.hasDot) this.hasDot = hasDot;
+        this.hasDot |= hasDot;
     }
 
     /**
      * Gets the width/height size ratio.
+     *
      * @return The width/height size ratio
      */
     public double getSizeRatio() {
@@ -78,7 +84,8 @@ public class TrainedCharacterData {
 
     /**
      * Adds the given width and height variables to the internal list to be put into calculations upon invoking
-     * {@link #finishRecalculations()}
+     * {@link #finishRecalculations()}.
+     *
      * @param width The width of the character
      * @param height The height of the character
      */
@@ -92,6 +99,7 @@ public class TrainedCharacterData {
     /**
      * Does the same thing as {@link #recalculateTo(double, double)} but with a {@link SearchCharacter}, and it
      * includes its percentages as well.
+     *
      * @param searchCharacter The {@link SearchCharacter} to be recalculated to
      */
     public void recalculateTo(SearchCharacter searchCharacter) {
@@ -112,26 +120,30 @@ public class TrainedCharacterData {
      */
     public void finishRecalculations() {
         OptionalDouble widthAverageOptional = recalculatingWidths.stream().mapToDouble(t -> t).average();
-        this.widthAverage = widthAverageOptional.isPresent() ? widthAverageOptional.getAsDouble() : 0D;
+        this.widthAverage = widthAverageOptional.orElse(0D);
 
         OptionalDouble heightAverageOptional = recalculatingHeights.stream().mapToDouble(t -> t).average();
-        this.heightAverage = heightAverageOptional.isPresent() ? heightAverageOptional.getAsDouble() : 0D;
+        this.heightAverage = heightAverageOptional.orElse(0D);
 
         this.sizeRatio = this.heightAverage != 0 ? this.widthAverage / this.heightAverage : 0;
 
-        if (value == ' ') return;
+        if (value == ' ') {
+            return;
+        }
 
         this.segmentPercentages = new double[8 + 9];
+        
         for (int i = 0; i < 8 + 9; i++) {
             int finalI = i;
-            OptionalDouble optionalDouble = recalculatingList.stream().mapToDouble(t -> t[finalI]).average();
-            this.segmentPercentages[i] = optionalDouble.isPresent() ? optionalDouble.getAsDouble() : 0;
+            this.segmentPercentages[i] = recalculatingList.stream().mapToDouble(t -> t[finalI]).average().orElse(0D);
         }
     }
 
     /**
      * If the given value is less than the minimum center, it's the new minimum center, and if it's bigger than the
-     * maximum center, it will be the new maximum center. (Retrievable via {@link #getMaxCenter() and {@link #getMinCenter()}})
+     * maximum center, it will be the new maximum center. (Retrievable via {@link #getMaxCenter() and {
+     * @link #getMinCenter()}}).
+     *
      * @param center The value to add as center
      */
     public void recalculateCenter(double center) {
@@ -139,13 +151,19 @@ public class TrainedCharacterData {
             minCenter = center;
             maxCenter = center;
         } else {
-            if (center > maxCenter) maxCenter = center;
-            if (center < minCenter) minCenter = center;
+            if (center > maxCenter) {
+                maxCenter = center;
+            }
+            
+            if (center < minCenter) {
+                minCenter = center;
+            }
         }
     }
 
     /**
      * Gets the average width for everything trained with this object.
+     *
      * @return The average width. Will return 0 if {@link #finishRecalculations()} has not been invoked.
      */
     public double getWidthAverage() {
@@ -154,6 +172,7 @@ public class TrainedCharacterData {
 
     /**
      * Gets the average height for everything trained with this object.
+     *
      * @return The average height. Will return 0 if {@link #finishRecalculations()} has not been invoked.
      */
     public double getHeightAverage() {
@@ -162,6 +181,7 @@ public class TrainedCharacterData {
 
     /**
      * Gets the minimum center value of all the training data used.
+     *
      * @return The minimum center value of all the training data used
      */
     public double getMinCenter() {
@@ -170,6 +190,7 @@ public class TrainedCharacterData {
 
     /**
      * Gets the maximum center value of all the training data used.
+     *
      * @return The maximum center value of all the training data used
      */
     public double getMaxCenter() {
@@ -179,14 +200,16 @@ public class TrainedCharacterData {
     /**
      * Gets if anything has been recalcuated/prepared to be recalculated to the character, e.g. by using
      * {@link #recalculateTo(double, double)} or {@link #recalculateTo(SearchCharacter)}.
+     *
      * @return If anything has been recalculated
      */
     public boolean isEmpty() {
-        return this.empty;
+        return empty;
     }
 
     /**
      * Gets the letter meta of the character.
+     *
      * @return The letter meta of the character
      */
     public LetterMeta getLetterMeta() {
@@ -195,6 +218,7 @@ public class TrainedCharacterData {
 
     /**
      * Sets the letter meta of the character.
+     *
      * @param letterMeta The letter meta of the character
      */
     public void setLetterMeta(LetterMeta letterMeta) {
@@ -205,4 +229,5 @@ public class TrainedCharacterData {
     public String toString() {
         return String.valueOf(value);
     }
+    
 }
